kas::widget::view: reorganise modules
