+++ src/widget/view/mod.rs
@@ -7,80 +7,16 @@
 //!
 //! View widgets exist as a view over some shared data.
 
-use super::Label;
-use kas::prelude::*;
-
 mod data_traits;
 mod filter;
-mod list;
-mod shared;
-mod single;
+mod list_view;
+mod shared_data;
+mod single_view;
+mod view_widget;
 
 pub use data_traits::{ListData, SingleData, SingleDataMut};
 pub use filter::{Filter, FilteredList, SimpleCaseInsensitiveFilter};
-pub use list::{ListMsg, ListView, SelectionMode};
-pub use shared::{SharedConst, SharedRc};
-pub use single::SingleView;
-
-/// View widgets
-///
-/// Implementors are able to view data of type `T`.
-pub trait ViewWidget<T>: Widget {
-    /// Construct a default instance (with no data)
-    fn default() -> Self
-    where
-        T: Default;
-    /// Construct an instance from a data value
-    fn new(data: T) -> Self;
-    /// Set the viewed data
-    fn set(&mut self, data: T) -> TkAction;
-}
-
-//TODO(spec): enable this as a specialisation of the T: ToString impl
-// In the mean-time we only lose the Markdown impl by disabling this
-/*
-impl<T: Clone + Default + FormattableText + 'static> ViewWidget<T> for Label<T> {
-    fn default() -> Self {
-        Label::new(T::default())
-    }
-    fn new(data: T) -> Self {
-        Label::new(data.clone())
-    }
-    fn set(&mut self, data: &T) -> TkAction {
-        self.set_text(data.clone())
-    }
-}
-*/
-
-impl<T: Default + ToString> ViewWidget<T> for Label<String> {
-    fn default() -> Self {
-        Label::new(T::default().to_string())
-    }
-    fn new(data: T) -> Self {
-        Label::new(data.to_string())
-    }
-    fn set(&mut self, data: T) -> TkAction {
-        self.set_text(data.to_string())
-    }
-}
-
-/// Default view assignments
-///
-/// This trait may be implemented to assign a default view widget to a specific
-/// data type.
-pub trait DefaultView: Sized {
-    type Widget: ViewWidget<Self>;
-}
-
-// TODO(spec): enable this over more specific implementations
-/*
-impl<T: Clone + Default + FormattableText + 'static> DefaultView for T {
-    type Widget = Label<T>;
-}
-*/
-impl DefaultView for String {
-    type Widget = Label<String>;
-}
-impl<'a> DefaultView for &'a str {
-    type Widget = Label<String>;
-}
+pub use list_view::{ListMsg, ListView, SelectionMode};
+pub use shared_data::{SharedConst, SharedRc};
+pub use single_view::SingleView;
+pub use view_widget::{DefaultView, ViewWidget};
