+++ src/widget/view/shared_data.rs
