+++ src/widget/view/single_view.rs
